# src/main/java/com/alfre/DHHotel/adapter/web/dto/AuthResponse.java
package com.alfre.DHHotel.adapter.web.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    String token;
}