# src/main/java/com/alfre/DHHotel/adapter/web/dto/LoginRequest.java
package com.alfre.DHHotel.adapter.web.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class LoginRequest {
    String username; // email
    String password;
}