# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Hotel Management API
  version: 1.0.0
  description: |
    API completa para la gestión integral de un sistema hotelero que incluye:
    - Autenticación y gestión de usuarios
    - Administración de habitaciones
    - Gestión de reservaciones y pagos
    - Administración de clientes y usuarios
  contact:
    name: Desarrollador de la API
    email: alfredo.sobrados.gonzalez@gmail.com
  license:
    name: License MIT
    url: https://mit-license.org/
servers:
  - url: http://localhost:8080/api
    description: Entorno de desarrollo local
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Iniciar sesión en el sistema
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Credenciales inválidas o formato incorrecto
        '404':
          description: Usuario no encontrado
  /auth/register/client:
    post:
      tags: [Authentication]
      summary: Registrar cliente
      operationId: registerClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestClient'
      responses:
        '200':
          description: Registro de cliente exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error en el registro del cliente
  /auth/register/admin:
    post:
      tags: [Authentication]
      summary: Registrar administrador
      operationId: registerAdministrator
      security:
        - bearerAuth: [superadmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestAdmin'
      responses:
        '200':
          description: Registro de administrador exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Error en el registro del administrador
  /auth/me/client:
    get:
      tags: [Authentication]
      summary: Obtener información del cliente autenticado
      operationId: getInfoClient
      security:
        - bearerAuth: [client]
      responses:
        '200':
          description: Información del cliente obtenida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientInfo'
        '500':
          description: Error en la autenticación o en el servicio
  /auth/me/admin:
    get:
      tags: [Authentication]
      summary: Obtener información del administrador autenticado
      operationId: getInfoAdmin
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Información del administrador obtenida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminInfo'
        '500':
          description: Error en la autenticación o en el servicio
  /auth/me:
    put:
      tags: [Authentication]
      summary: Actualizar perfil del cliente
      operationId: updateProfile
      security:
        - bearerAuth: [client]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Perfil actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientInfo'
        '400':
          description: Error en la solicitud (datos inválidos)
        '404':
          description: Cliente no registrado
  /auth/change-password:
    put:
      tags: [Authentication]
      summary: Cambiar la contraseña del usuario
      operationId: changePassword
      security:
        - bearerAuth: [admin, client]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Contraseña cambiada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '500':
          description: Error en la autenticación o en el servicio
  /auth/change-email:
    put:
      tags: [Authentication]
      summary: Cambiar el correo electrónico del usuario
      operationId: changeEmail
      security:
        - bearerAuth: [admin, client]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailRequest'
      responses:
        '200':
          description: Correo electrónico actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeEmailResponse'
        '400':
          description: Error en la solicitud (datos inválidos)
        '500':
          description: Error en la autenticación o en el servicio

  # ==============================================
  # Administrators Endpoints
  # ==============================================
  /superadmin/admins:
    get:
      tags: [Administrators]
      summary: Obtener todos los administradores
      operationId: getAllAdministrators
      security:
        - bearerAuth: [superadmin]
      responses:
        '200':
          description: Lista de administradores obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Administrator'
        '404':
          description: No hay administradores registrados
        '500':
          description: Error interno del servidor
  /superadmin/admin/userId/{userId}:
    get:
      tags: [Administrators]
      summary: Obtener administrador por userId
      operationId: getAdministratorByUserId
      security:
        - bearerAuth: [superadmin]
      parameters:
        - name: userId
          in: path
          required: true
          description: Identificador del usuario (userId)
          schema:
            type: integer
            example: 1001
      responses:
        '200':
          description: Administrador encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Administrator'
        '404':
          description: El administrador solicitado no existe
        '500':
          description: Error interno del servidor
  /superadmin/admin/id/{id}:
    get:
      tags: [Administrators]
      summary: Obtener administrador por ID
      operationId: getAdministratorById
      security:
        - bearerAuth: [superadmin]
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del administrador
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Administrador encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Administrator'
        '404':
          description: El administrador solicitado no existe
        '500':
          description: Error interno del servidor
  /superadmin/admin/{userId}:
    put:
      tags: [Administrators]
      summary: Actualizar datos de un administrador
      operationId: updateAdministrator
      security:
        - bearerAuth: [superadmin]
      parameters:
        - name: userId
          in: path
          required: true
          description: userId del administrador a actualizar
          schema:
            type: integer
            example: 1001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Administrator'
      responses:
        '200':
          description: Actualización exitosa
        '400':
          description: No se ha podido actualizar (datos inválidos o inexistentes)
        '500':
          description: Error interno del servidor
  /superadmin/admin/{id}:
    delete:
      tags: [Administrators]
      summary: Eliminar un administrador por ID
      operationId: deleteAdministrator
      security:
        - bearerAuth: [superadmin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del administrador a eliminar
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Administrador eliminado correctamente
        '400':
          description: No se ha podido eliminar (ID inválido o inexistente)
        '500':
          description: Error interno del servidor

  # ==============================================
  # Clients Endpoints
  # ==============================================
  /admin/clients:
    get:
      tags:
        - Clients
      summary: Listar todos los clientes
      operationId: getAllClients
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Lista completa de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '404':
          description: No hay clientes registrados
        '500':
          description: Error interno del servidor
  /admin/client/{id}:
    get:
      tags: [Clients]
      summary: Obtener cliente por ID
      operationId: getClientById
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Detalles del cliente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '404':
          description: El cliente solicitado no existe
        '500':
          description: Error interno del servidor

    delete:
      tags: [Clients]
      summary: Eliminar cliente por ID
      operationId: deleteClient
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID del cliente a eliminar
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Cliente eliminado correctamente
        '404':
          description: El cliente no ha sido eliminado (no existe)
        '500':
          description: Error interno del servidor

  # ==============================================
  # Rooms Endpoints
  # ==============================================
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: usuario@dhhotel.com
        password:
          type: string
          format: password
          example: SecurePass123!
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    RegisterRequestAdmin:
      type: object
      properties:
        username:
          type: string
          format: email
          example: usuario@dhhotel.com
        password:
          type: string
          format: password
          example: SecurePass123!
        role:
          type: string
          enum: [SUPERADMIN, ADMIN, CLIENT]
          example: CLIENT
        name:
          type: string
          example: Domingo Suárez
      required:
        - username
        - email
        - role
        - password
    RegisterRequestClient:
      type: object
      properties:
        username:
          type: string
          format: email
          example: usuario@dhhotel.com
        password:
          type: string
          format: password
          example: SecurePass123!
        role:
          type: string
          enum: [SUPERADMIN, ADMIN, CLIENT]
          example: CLIENT
        first_name:
          type: string
          example: Roberto
        last_name:
          type: string
          example: Pérez
        phone:
          type: string
          format: phone
          example: +34 612 234 435
      required:
        - username
        - password
        - first_name
        - last_name
        - phone
        - role
    ClientInfo:
      type: object
      properties:
        email:
          type: string
          example: ramon@dhhotel.com
        first_name:
          type: string
          example: Ramón
        last_name:
          type: string
          example: Álvarez
        phone:
          type: string
          example: +34 654 123 456
      required:
        - email
        - first_name
        - last_name
        - phone
    AdminInfo:
      type: object
      properties:
        email:
          type: string
          example: julian@dhhotel.com
        name:
          type: string
          example: Julián Lago
        role:
          type: string
          enum: [ADMIN, SUPERADMIN]
          example: ADMIN
      required:
        - email
        - name
        - role
    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          example: Joaquín
        last_name:
          type: string
          example: García
        email:
          type: string
          example: joaquin@dhhotel.com
      required:
        - first_name
        - last_name
        - email
    ChangePasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
          example: newPasswordSecure12!
      required:
        - newPassword
    ChangePasswordResponse:
      type: object
      properties:
        response:
          type: string
          example: Cambio de email CORRECTO.
      required:
        - response
    ChangeEmailRequest:
      type: object
      properties:
        newEmail:
          type: string
          example: newEmail@dhhotel.com
      required:
        - newEmail
    ChangeEmailResponse:
      type: object
      properties:
        response:
          type: string
          example: Cambio de contraseña CORRECTO.
      required:
        - response

    # Clients
    Client:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 2
        first_name:
          type: string
          example: "Juan"
        last_name:
          type: string
          example: "Pérez"
        phone:
          type: string
          example: +34 687 123 456
      required:
        - id
        - user_id
        - first_name
        - last_name
        - phone

    # Users
...
    Administrator:
      type: object
      properties:
        id:
          type: integer
          example: 5
        userId:
          type: integer
          example: 6
        name:
          type: string
          example: "Álvaro Sanz"
      required:
        - id
        - userId
        - name

    # Rooms
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []